session: move all Bolt DB code to `kvdb_store.go`
Keep all KVDB logic contained within the same file.
